climber servo fix (#16)
The servo "remembers" its instructions, even after the bot is rebooted.  So in case its latest instructions were to go to release position, then the bot was turned off, then the servo was manually set to its rest position, it doesn't boot up and go back to its release position.